move swap struct to swap.rs
