+++ smart-contracts/osmosis/contracts/cl-vault/src/msg.rs
@@ -8,7 +8,10 @@ use crate::query::{
     AssetsBalanceResponse, PoolResponse, PositionResponse, RangeAdminResponse,
     UserSharesBalanceResponse, VerifyTickCacheResponse,
 };
-use crate::state::{Metadata, VaultConfig};
+use crate::{
+    state::{Metadata, VaultConfig},
+    vault::swap::SwapOperation,
+};
 
 /// Extension execute messages for an apollo autocompounding vault
 #[cw_serde]
@@ -94,16 +97,6 @@ pub struct ModifyRangeMsg {
 #[cw_serde]
 pub struct MergePositionMsg {
     pub position_ids: Vec<u64>,
-}
-
-// struct used by swap.rs on swap non vault funds
-#[cw_serde]
-pub struct SwapOperation {
-    pub token_in_denom: String,
-    pub pool_id_0: u64, // the osmosis pool_id as mandatory to have at least the chance to swap on CL pools
-    pub pool_id_1: u64, // the osmosis pool_id as mandatory to have at least the chance to swap on CL pools
-    pub forced_swap_route_token_0: Option<Vec<SwapAmountInRoute>>,
-    pub forced_swap_route_token_1: Option<Vec<SwapAmountInRoute>>,
 }
 
 /// Extension query messages for an apollo autocompounding vault
